Add atlas page property to sprite manager
Atlas page property to the sprite manager UI and data model. If a sprite is very large and you want to explicitly specify it should be on its own atlas page, that's what this would be for. Maybe a less rudimentary approach would be better but this works for my specific use case